Adjust the source code and build enviornment so that Mangos Zero will build on ARM32. (#79)
This was primarily done to test the feasibility of running on a RPi4 running
Debian Buster (it works quite well).

Summary of changes:

  - Reworked patches I found on the Mangos forum, original pastebins here:
      https://pastebin.com/BxqCmCML
  - Adjusted cmake's arch detection to include ARM32/ARM64
  - Adjusted compiler flags for Linux builds and supressed some extraneous notes

I have played for about an hour and noticed no issues or differences from i386/x86_64 VMs.

One note, currently the map extraction tools for ARM are broken.  They compile but do
work properly.  I'll work on that next.