Update Player.cpp for Issue "All spells buyable from class trainer"
"All spells buyable from class trainer"
The issue seems to stem from the conditions in the code not properly checking the player's level against the required level (reqLevel) of the spell at the trainers. The code snippet as it stands checks if the spell is a profession spell and if the required level of the spell is less than reqLevel, but it doesn't compare it to the player's actual level, which it seems like it should.